Refactor DateOnlyExtensions into Core project
