+++ OpenBudgeteer.Blazor/Pages/Transaction.razor.cs
@@ -8,6 +8,7 @@ using OpenBudgeteer.Blazor.Common;
 using OpenBudgeteer.Blazor.Common.CustomMudFilter;
 using OpenBudgeteer.Blazor.Shared.Dialog;
 using OpenBudgeteer.Core.Common;
+using OpenBudgeteer.Core.Common.Extensions;
 using OpenBudgeteer.Core.Data.Contracts.Services;
 using OpenBudgeteer.Core.ViewModels.EntityViewModels;
 using OpenBudgeteer.Core.ViewModels.Helper;
