+++ OpenBudgeteer.Blazor/Shared/Dialog/BucketDetailsDialog.razor
@@ -4,6 +4,7 @@
 @using System.Globalization
 @using OpenBudgeteer.Blazor.Common.CustomMudFilter
 @using OpenBudgeteer.Core.Common
+@using OpenBudgeteer.Core.Common.Extensions
 @using OpenBudgeteer.Core.ViewModels.EntityViewModels
 @using Color = MudBlazor.Color
 @using Size = MudBlazor.Size
