+++ OpenBudgeteer.Blazor/Shared/Dialog/TransactionDialog.razor
@@ -1,6 +1,7 @@
 ﻿@using System.Globalization
 @using OpenBudgeteer.Blazor.Common
 @using OpenBudgeteer.Blazor.Common.CustomMudFilter
+@using OpenBudgeteer.Core.Common.Extensions
 @using OpenBudgeteer.Core.Data.Entities.Models
 @using OpenBudgeteer.Core.ViewModels.EntityViewModels
 @using OpenBudgeteer.Core.ViewModels.Helper
